Adjust time calculation in HelperFunctions
The calculations for time in seconds have been modified in the HelperFunctions utility. If the total seconds value is less than 1, it will now return 1 instead of the original decimal. This adjustment ensures a minimum value of 1 second in certain scenarios, preventing possible inaccuracies or errors in the time display.